+++ src/components/flat-table/flat-table-cell/flat-table-cell.style.js
@@ -2,6 +2,13 @@ import styled, { css } from "styled-components";
 import { padding } from "styled-system";
 
 import baseTheme from "../../../style/themes/base";
+import { toColor } from "../../../style/utils/color";
+
+const verticalBorderSizes = {
+  small: "1px",
+  medium: "2px",
+  large: "4px",
+};
 
 const StyledFlatTableCell = styled.td`
   ${({
@@ -13,6 +20,8 @@ const StyledFlatTableCell = styled.td`
     colWidth,
     isTruncated,
     expandable,
+    verticalBorder,
+    verticalBorderColor,
   }) => css`
     background-color: #fff;
     border-width: 0;
@@ -44,6 +53,17 @@ const StyledFlatTableCell = styled.td`
  
         ${padding}
       }
+
+      ${verticalBorder &&
+      css`
+        border-right: ${verticalBorderSizes[verticalBorder]} solid
+          ${theme.table.secondary};
+      `}
+
+      ${verticalBorderColor &&
+      css`
+        border-right-color: ${toColor(theme, verticalBorderColor)};
+      `}
     }
 
     &:first-of-type {