+++ src/components/flat-table/flat-table.style.js
@@ -110,6 +110,7 @@ const StyledFlatTableWrapper = styled.div`
           ${StyledFlatTableHead} ${StyledFlatTableCheckbox} {
             background-color: ${theme.flatTable.light.headerBackground};
             border-right: 1px solid ${theme.flatTable.light.border};
+            border-bottom-color: ${theme.flatTable.light.border};
           }
         `;
 
@@ -121,6 +122,7 @@ const StyledFlatTableWrapper = styled.div`
             background-color: ${theme.flatTable.transparentBase
               .headerBackground};
             border-right: 1px solid ${theme.flatTable.transparentBase.border};
+            border-bottom-color: ${theme.table.secondary};
           }
         `;
 
@@ -132,6 +134,7 @@ const StyledFlatTableWrapper = styled.div`
             background-color: ${theme.flatTable.transparentWhite
               .headerBackground};
             border-right: 1px solid ${theme.flatTable.transparentWhite.border};
+            border-bottom-color: ${theme.table.secondary};
           }
         `;
       // default theme is "dark"
@@ -143,6 +146,7 @@ const StyledFlatTableWrapper = styled.div`
             background-color: ${theme.flatTable.dark.headerBackground};
             border-right: 1px solid ${theme.flatTable.dark.border};
             color: ${theme.colors.white};
+            border-bottom-color: ${theme.flatTable.dark.border};
           }
         `;
     }
@@ -157,6 +161,7 @@ const StyledFlatTableWrapper = styled.div`
         border-right: 2px solid
           ${theme.flatTable.drawerSidebar.headerBackground};
         color: ${theme.colors.black};
+        border-bottom-color: ${theme.table.secondary};
       }
     `}