+++ src/style/themes/base/base-theme.config.js
@@ -160,6 +160,7 @@ export default (palette) => {
 
       hover: palette.slateTint(95),
       headerIconColor: palette.slateTint(60),
+      headerAlternativeBackground: palette.slateTint(10),
       selected: palette.slateTint(85),
       highlighted: palette.slateTint(90),
     },
@@ -342,7 +343,8 @@ export default (palette) => {
     },
 
     switch: {
-      off: "#CCD6DB",
+      off: palette.slateTint(80),
+      offHover: palette.slateTint(60),
     },
 
     tile: {